+++ packages_flutter/sqflite_test_app_darwin_pod/lib/main.dart
@@ -1,0 +1,10 @@
+import 'package:sqflite_example/main.dart' as impl;
+
+Future<void> main() async {
+  await mainSqfliteTestApp();
+}
+
+/// Main sqflite test app.
+Future<void> mainSqfliteTestApp() async {
+  impl.main();
+}