+++ packages_flutter/sqflite_test_app_darwin_pod/tool/flutter_run_macos.dart
@@ -1,0 +1,7 @@
+import 'package:sqflite_support/spm.dart';
+import 'package:sqflite_support/test_project.dart';
+
+Future<void> main() async {
+  await disableSpm();
+  await runMacOS();
+}