+++ sqflite_common/lib/src/logger/sqflite_logger.dart
@@ -10,7 +10,6 @@ import 'package:sqflite_common/src/database_mixin.dart';
 import 'package:sqflite_common/src/env_utils.dart';
 import 'package:sqflite_common/src/factory.dart';
 import 'package:sqflite_common/src/factory_mixin.dart';
-import 'package:sqflite_common/src/sql_command.dart';
 import 'package:sqflite_common/src/transaction.dart';
 
 /// Log helper to avoid overflow