+++ sqflite_darwin/darwin/Classes/SqfliteCursor.m
@@ -1,0 +1,10 @@
+#import "SqfliteCursor.h"
+#import "SqfliteDarwinImport.h"
+
+@implementation SqfliteCursor
+
+@synthesize cursorId;
+@synthesize pageSize;
+@synthesize resultSet;
+
+@end