+++ sqflite_platform_interface/example/lib/database/database.dart
@@ -1,0 +1,2 @@
+export 'database_impl.dart';
+export 'database_io.dart' if (dart.library.js_interop) 'database_web.dart';