+++ sqflite_platform_interface/lib/sqflite_platform_interface.dart
@@ -1,3 +1,4 @@
+import 'package:meta/meta.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'package:sqflite_common/sqflite.dart' as impl;
 import 'package:sqflite_common/sqlite_api.dart';
@@ -11,7 +12,7 @@ class SqflitePlatform extends PlatformInterface {
 
   /// Method channel database factory.
   /// Set this factory during register for sqflite_android and sqflite_darwin
-  DatabaseFactory get databaseFactoryMethodChannel =>
+  static DatabaseFactory get databaseFactoryMethodChannel =>
       impl.sqfliteDatabaseFactoryDefault;
 
   /// Get the database factory
@@ -19,6 +20,18 @@ class SqflitePlatform extends PlatformInterface {
 
   /// Platform specific plugins should set this with their own platform-specific
   set databaseFactory(DatabaseFactory databaseFactory) {
+    impl.databaseFactory = databaseFactory;
+  }
+
+  /// Get the database factory or null if not set
+  @protected
+  static DatabaseFactory? get databaseFactoryOrNull =>
+      // ignore: invalid_use_of_visible_for_testing_member
+      impl.databaseFactoryOrNull;
+
+  /// Platform specific plugins should set this with their own platform-specific
+  @protected
+  static set databaseFactoryOrNull(DatabaseFactory? databaseFactory) {
     impl.databaseFactory = databaseFactory;
   }
 }