+++ sqflite_test_app/test_driver/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_driver/driver_extension.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sqflite/sqflite.dart';
+import 'package:sqflite_common/sqflite_dev.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 import 'package:sqflite_common_test/all_test.dart' as all;
 import 'package:sqflite_common_test/sqflite_test.dart';
@@ -32,7 +33,9 @@ void main() {
     group('driver with 2 threads', () {
       setUpAll(() async {
         // ignore: deprecated_member_use, deprecated_member_use_from_same_package
-        await Sqflite.devSetOptions(SqfliteOptions()..androidThreadCount = 2);
+        await databaseFactory
+            // ignore: deprecated_member_use
+            .setOptions(SqfliteOptions()..androidThreadCount = 2);
       });
       all.run(testContext);
     });