+++ sqflite_platform_interface/CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 2.4.0-0
 
-* Initial revision
+* Initial implementation from sqflite v2.3.3+2